requirements.txt
Django
djangorestframework
psycopg2-binary
requests
python-dotenv
pytest